requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
asgiref==3.8.1
attrs==23.2.0
av==12.0.0
boto3==1.34.86
botocore==1.34.86
celluloid==0.2.0
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.29.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.10.0
Django==4.0.10
django-cors-headers==4.3.1
djangorestframework==3.15.1
easing-functions==1.0.4
exceptiongroup==1.2.0
feat==1.0.5
filelock==3.13.4
fire==0.6.0
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
grpcio==1.62.1
h5py==3.11.0
huggingface-hub==0.22.2
idna==2.10
imageio==2.34.0
importlib-metadata==7.1.0
importlib-resources==6.4.0
iniconfig==2.0.0
jax==0.4.26
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
kiwisolver==1.4.5
kornia==0.7.2
kornia-rs==0.1.3
lazy-loader==0.4
loguru==0.7.2
lxml==5.2.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
mediapipe==0.10.11
ml-dtypes==0.4.0
mpmath==1.3.0
networkx==3.2.1
nibabel==5.2.1
nilearn==0.10.4
nltools==0.5.1
numexpr==2.8.4
numpy==1.23.5
omegaconf==2.3.0
opencv-contrib-python==4.9.0.80
opencv-python==4.7.0.72
opencv-python-headless==4.8.0.74
opt-einsum==3.3.0
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pluggy==1.4.0
protobuf==3.20.3
psutil==5.9.8
py-cpuinfo==9.0.0
py-feat==0.6.2
pybboxes==0.1.6
pycparser==2.22
pynv==0.3
pyparsing==3.1.2
pypose==0.6.7
pytest==7.4.4
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.1
pywavelets==1.6.0
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
roboflow==1.1.27
s3transfer==0.10.1
safetensors==0.4.3
sahi==0.11.15
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
shapely==2.0.4
six==1.16.0
smmap==5.0.1
sounddevice==0.4.6
sqlparse==0.4.4
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.4.0
tifffile==2024.2.12
timm==0.9.16
tomli==2.0.1
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typing-extensions==4.10.0
tzdata==2024.1
ultralytics==8.2.1
urllib3==1.26.18
werkzeug==3.0.2
win32-setctime==1.1.0
xgboost==2.0.3
yolov5==7.0.13
yolov7==0.0.1
zipp==3.18.1
zope.interface==6.3
